setting interests now works

+++ backend_app/routes/interests.js
@@ -10,7 +10,7 @@ export default {
 
     if (_.isEmpty(userId)) {
       throw new ApplicationError(
-        'Missing required fields', 400, { missing: ['id'], requiredFields }
+        'Missing required fields', 400, { missing: ['userId'], requiredFields }
       );
     }
 
@@ -23,13 +23,21 @@ export default {
 
   add: (req, res, next) => {
     const requiredFields = ['userId', 'topicId'];
-    const values = req.body;
+    const values = Object.assign({}, req.body, req.params);
 
-    if (_.compact(_.values(values)).length !== requiredFields.length) {
+    if (_.compact(_.values(_.pick(values, requiredFields))).length !== requiredFields.length) {
       throw new ApplicationError('Missing required fields', 400, { requiredFields });
     }
 
-    const dbFunc = req.body.primary ? interestsDb.add : interestsDb.updatePrimary;
+    if (isNaN(values.userId)) {
+      throw new ApplicationError('userId must be an integer', 400);
+    }
+
+    if (res.locals.authData.id !== parseInt(values.userId, 10)) {
+      throw new ApplicationError('You may only add interests for your own account', 403);
+    }
+
+    const dbFunc = req.body.primary ? interestsDb.updatePrimary : interestsDb.add;
     dbFunc(values.userId, values.topicId).then((result) => {
       if (_.isEmpty(result)) {
         res.status(200).send('Already added');
@@ -40,10 +48,22 @@ export default {
 
   remove: (req, res, next) => {
     const requiredFields = ['userId', 'topicId'];
-    const values = req.body;
+    const values = Object.assign({}, req.body, req.params);
 
     if (_.compact(_.values(values)).length !== requiredFields.length) {
       throw new ApplicationError('Missing required fields', 400, { requiredFields });
+    }
+
+    if (isNaN(values.userId)) {
+      throw new ApplicationError('userId must be an integer', 400);
+    }
+
+    if (isNaN(values.topicId)) {
+      throw new ApplicationError('topicId must be an integer', 400);
+    }
+
+    if (res.locals.authData.id !== parseInt(values.userId, 10)) {
+      throw new ApplicationError('You may only remove interests for your own account', 403);
     }
 
     interestsDb.remove(values.userId, values.topicId).then(() => {
